HBASE-24968 : Move META_COMPARATOR to subclass MetaCellComparator
