+++ hbase-common/src/main/java/org/apache/hadoop/hbase/CellComparatorImpl.java
@@ -17,16 +17,12 @@
  */
 package org.apache.hadoop.hbase;
 
-import java.nio.ByteBuffer;
 import java.util.Comparator;
 import org.apache.hadoop.hbase.KeyValue.Type;
 import org.apache.hadoop.hbase.util.ByteBufferUtils;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.yetus.audience.InterfaceAudience;
 import org.apache.yetus.audience.InterfaceStability;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.apache.hbase.thirdparty.com.google.common.primitives.Longs;
 
 /**
  * Compare two HBase cells.  Do not use this method comparing <code>-ROOT-</code> or
@@ -48,19 +44,12 @@ import org.apache.hbase.thirdparty.com.google.common.primitives.Longs;
 @InterfaceAudience.Private
 @InterfaceStability.Evolving
 public class CellComparatorImpl implements CellComparator {
-  static final Logger LOG = LoggerFactory.getLogger(CellComparatorImpl.class);
 
   /**
    * Comparator for plain key/values; i.e. non-catalog table key/values. Works on Key portion
    * of KeyValue only.
    */
   public static final CellComparatorImpl COMPARATOR = new CellComparatorImpl();
-
-  /**
-   * A {@link CellComparatorImpl} for <code>hbase:meta</code> catalog table
-   * {@link KeyValue}s.
-   */
-  public static final CellComparatorImpl META_COMPARATOR = new MetaCellComparator();
 
   @Override
   public final int compare(final Cell a, final Cell b) {
@@ -213,7 +202,7 @@ public class CellComparatorImpl implements CellComparator {
   /**
    * Compares the row part of the cell with a simple plain byte[] like the
    * stopRow in Scan. This should be used with context where for hbase:meta
-   * cells the {{@link #META_COMPARATOR} should be used
+   * cells the {{@link MetaCellComparator#META_COMPARATOR} should be used
    *
    * @param left
    *          the cell to be compared
@@ -291,117 +280,6 @@ public class CellComparatorImpl implements CellComparator {
     return Long.compare(rtimestamp, ltimestamp);
   }
 
-  /**
-   * A {@link CellComparatorImpl} for <code>hbase:meta</code> catalog table
-   * {@link KeyValue}s.
-   */
-  public static class MetaCellComparator extends CellComparatorImpl {
-    // TODO: Do we need a ByteBufferKeyValue version of this?
-    @Override
-    public int compareRows(final Cell left, final Cell right) {
-      return compareRows(left.getRowArray(), left.getRowOffset(), left.getRowLength(),
-          right.getRowArray(), right.getRowOffset(), right.getRowLength());
-    }
-
-    @Override
-    public int compareRows(Cell left, byte[] right, int roffset, int rlength) {
-      return compareRows(left.getRowArray(), left.getRowOffset(), left.getRowLength(), right,
-          roffset, rlength);
-    }
-
-    @Override
-    public int compare(final Cell a, final Cell b, boolean ignoreSequenceid) {
-      int diff = compareRows(a, b);
-      if (diff != 0) {
-        return diff;
-      }
-
-      diff = compareWithoutRow(a, b);
-      if (diff != 0) {
-        return diff;
-      }
-
-      // Negate following comparisons so later edits show up first mvccVersion: later sorts first
-      return ignoreSequenceid? diff: Longs.compare(b.getSequenceId(), a.getSequenceId());
-    }
-
-    private static int compareRows(byte[] left, int loffset, int llength, byte[] right, int roffset,
-        int rlength) {
-      int leftDelimiter = Bytes.searchDelimiterIndex(left, loffset, llength, HConstants.DELIMITER);
-      int rightDelimiter = Bytes
-          .searchDelimiterIndex(right, roffset, rlength, HConstants.DELIMITER);
-      // Compare up to the delimiter
-      int lpart = (leftDelimiter < 0 ? llength : leftDelimiter - loffset);
-      int rpart = (rightDelimiter < 0 ? rlength : rightDelimiter - roffset);
-      int result = Bytes.compareTo(left, loffset, lpart, right, roffset, rpart);
-      if (result != 0) {
-        return result;
-      } else {
-        if (leftDelimiter < 0 && rightDelimiter >= 0) {
-          return -1;
-        } else if (rightDelimiter < 0 && leftDelimiter >= 0) {
-          return 1;
-        } else if (leftDelimiter < 0) {
-          return 0;
-        }
-      }
-      // Compare middle bit of the row.
-      // Move past delimiter
-      leftDelimiter++;
-      rightDelimiter++;
-      int leftFarDelimiter = Bytes.searchDelimiterIndexInReverse(left, leftDelimiter, llength
-          - (leftDelimiter - loffset), HConstants.DELIMITER);
-      int rightFarDelimiter = Bytes.searchDelimiterIndexInReverse(right, rightDelimiter, rlength
-          - (rightDelimiter - roffset), HConstants.DELIMITER);
-      // Now compare middlesection of row.
-      lpart = (leftFarDelimiter < 0 ? llength + loffset : leftFarDelimiter) - leftDelimiter;
-      rpart = (rightFarDelimiter < 0 ? rlength + roffset : rightFarDelimiter) - rightDelimiter;
-      result = Bytes.compareTo(left, leftDelimiter, lpart, right, rightDelimiter, rpart);
-      if (result != 0) {
-        return result;
-      } else {
-        if (leftDelimiter < 0 && rightDelimiter >= 0) {
-          return -1;
-        } else if (rightDelimiter < 0 && leftDelimiter >= 0) {
-          return 1;
-        } else if (leftDelimiter < 0) {
-          return 0;
-        }
-      }
-      // Compare last part of row, the rowid.
-      leftFarDelimiter++;
-      rightFarDelimiter++;
-      result = Bytes.compareTo(left, leftFarDelimiter, llength - (leftFarDelimiter - loffset),
-          right, rightFarDelimiter, rlength - (rightFarDelimiter - roffset));
-      return result;
-    }
-
-    @Override
-    public int compareRows(ByteBuffer row, Cell cell) {
-      byte [] array;
-      int offset;
-      int len = row.remaining();
-      if (row.hasArray()) {
-        array = row.array();
-        offset = row.position() + row.arrayOffset();
-      } else {
-        // We copy the row array if offheap just so we can do a compare. We do this elsewhere too
-        // in BBUtils when Cell is backed by an offheap ByteBuffer. Needs fixing so no copy. TODO.
-        array = new byte[len];
-        offset = 0;
-        ByteBufferUtils.copyFromBufferToArray(array, row, row.position(),
-          0, len);
-      }
-      // Reverse result since we swap the order of the params we pass below.
-      return -compareRows(cell, array, offset, len);
-    }
-
-    @Override
-    public Comparator getSimpleComparator() {
-      return this;
-    }
-  }
-
   @Override
   public Comparator getSimpleComparator() {
     return new BBKVComparator(this);
@@ -424,6 +302,6 @@ public class CellComparatorImpl implements CellComparator {
   public static CellComparator getCellComparator(byte [] tableName) {
     // FYI, TableName.toBytes does not create an array; just returns existing array pointer.
     return Bytes.equals(tableName, TableName.META_TABLE_NAME.toBytes())?
-      CellComparatorImpl.META_COMPARATOR: CellComparatorImpl.COMPARATOR;
+      MetaCellComparator.META_COMPARATOR: CellComparatorImpl.COMPARATOR;
   }
 }
