+++ hbase-server/src/test/java/org/apache/hadoop/hbase/MetaMockingUtil.java
@@ -100,7 +100,7 @@ public class MetaMockingUtil {
     }
 
     //important: sort the kvs so that binary search work
-    Collections.sort(kvs, CellComparatorImpl.META_COMPARATOR);
+    Collections.sort(kvs, MetaCellComparator.META_COMPARATOR);
 
     return Result.create(kvs);
   }
